Switch ByteRleReader to emit i8 instead of u8
Most cases we want i8 (For Int8Array, Union types) so we have to cast
from u8 to i8 which isn't ergonomic. Only cases we need u8 is for
boolean, but that can probably be refactored later.

+++ src/encoding/boolean.rs
@@ -63,7 +63,7 @@ impl<R: Read> Iterator for BooleanIter<R> {
         if self.bits_in_data == 0 {
             match self.iter.next() {
                 Some(Ok(data)) => {
-                    self.data = data;
+                    self.data = data as u8;
                     self.bits_in_data = 8;
                     Some(Ok(self.value()))
                 }
